+++ examples/2/main.js
@@ -1,33 +1,35 @@
 yamvc.$onReady(function () {
     "use strict";
 
-
-    var model, bar, ctr;
-
-    model = new yamvc.Model({
-        config: {namespace: 'likes'},
-        data: {
-            count: 0
-        }
-    });
-
-    bar = new app.view.Bar({
-        config: {
-            autoCreate: true,
-            models: {
-                likes: model
-            },
-            id: 'bar',
-            tpl: 'tpl-bar',
-            renderTo: '#container'
-        }
-    });
-
-    ctr = new yamvc.Controller({
+    new yamvc.Controller({
         config: {
             name: 'Main',
             views: {
-                topBar: yamvc.ViewManager.get('bar')
+                topBar: new app.view.Bar({
+                    config: {
+                        autoCreate: true,
+                        models: {
+                            likes: new yamvc.Model({
+                                config: {namespace: 'likes'},
+                                data: {
+                                    count: 0
+                                }
+                            })
+                        },
+                        id: 'bar',
+                        tpl: new yamvc.view.Template({
+                            config : {
+                                id: 'tpl-bar',
+                                tpl: [
+                                    '<div class="bar">',
+                                    'Count : {{likes.count}}',
+                                    '</div>'
+                                ]
+                            }
+                        }),
+                        renderTo: '#container'
+                    }
+                })
             },
             events: {
             }