+++ src/view/Template.js
@@ -1,13 +1,14 @@
 (function (window, undefined) {
     "use strict";
-    var yamvc = window.yamvc || {},
-        Proxy;
 
-    Proxy = yamvc.Core.$extend({
+    var yamvc = window.yamvc || {},
+        Template;
+
+    Template = yamvc.Core.$extend({
         init: function (opts) {
             var me = this, config;
 
-            Proxy.Parent.init.apply(this, arguments);
+            Template.Parent.init.apply(this, arguments);
 
             opts = opts || {};
             config = yamvc.$merge(me._config, opts.config);
@@ -16,75 +17,39 @@
             me.set('config', config);
 
             me.initConfig();
+            me.initTpl();
 
         },
-        read: function (action) {
+        initConfig: function () {
             var me = this,
-                opts,
-                id;
+                config = me.get('config');
 
-            if (!(action instanceof yamvc.data.Action))
-                throw new Error('yamvc.data.Proxy: read argument action should be instance of yamvc.data.Action');
+            Template.Parent.initConfig.apply(this);
 
-            opts = action.getOptions();
-            id = opts.params && opts.params.id;
+            if (!config.id)
+                throw new Error("yamvc.data.Template: Template need to have id");
 
-            if (!action.getOption('namespace'))
-                throw new Error('yamvc.data.Proxy: namespace should be set');
+            return me;
+        },
+        initTpl: function () {
+            var me = this,
+                html = me.getTpl(),
+                tpl = document.createElement('div');
 
-            if (typeof id === 'undefined') {
-                me.readBy(action);
-            } else {
-                me.readById(action);
+            if (Array.isArray(html)) {
+                html = html.join("");
             }
 
-            return me;
+            tpl.innerHTML = html;
+
+            me.set('html', tpl);
         },
-        create: function (action) {
-            var me = this;
-
-            if (!(action instanceof yamvc.data.Action))
-                throw new Error('yamvc.data.Proxy: create argument action should be instance of yamvc.data.Action');
-
-            if (!action.getOption('namespace'))
-                throw new Error('yamvc.data.Proxy: namespace should be set');
-
-            if (!action.getData() || typeof action.getData() !== 'object')
-                throw new Error('yamvc.data.Proxy: Data should be object');
-
-            return me;
-        },
-        update: function (action) {
-            var me = this;
-
-            if (!(action instanceof yamvc.data.Action))
-                throw new Error('yamvc.data.Proxy: update argument action should be instance of yamvc.data.Action');
-
-            if (!action.getOption('namespace'))
-                throw new Error('yamvc.data.Proxy: namespace should be set');
-
-            if (!action.getData() || typeof action.getData() !== 'object')
-                throw new Error('yamvc.data.Proxy: Data should be object');
-
-            return me;
-        },
-        destroy: function (action) {
-            var me = this;
-
-            if (!(action instanceof yamvc.data.Action))
-                throw new Error('yamvc.data.Proxy: destroy argument action should be instance of yamvc.data.Action');
-
-            if (!action.getOption('namespace'))
-                throw new Error('yamvc.data.Proxy: namespace should be set');
-
-            if (!action.getData() || typeof action.getData() !== 'object')
-                throw new Error('Data should be pass as object');
-
-            return me;
+        getHtml : function () {
+            return this._html;
         }
     });
 
     window.yamvc = yamvc;
-    window.yamvc.data = window.yamvc.data || {};
-    window.yamvc.data.Proxy = Proxy;
+    window.yamvc.view = window.yamvc.view || {};
+    window.yamvc.view.Template = Template;
 }(window));